Refatoramento das classes e inicio de tratamento de erro

+++ controle-familia-backend/src/main/java/com/example/controlefamiliabackend/repositories/UsuarioRepository.java
@@ -4,10 +4,11 @@ import com.example.controlefamiliabackend.models.UsuarioModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<UsuarioModel, Integer> {
+public interface UsuarioRepository extends JpaRepository<UsuarioModel, BigInteger> {
    boolean existsByDsEmail(String dsEmail);
 }
